+++ src/utils.rs
@@ -1,8 +1,12 @@
 use discord_rich_presence::{DiscordIpc, DiscordIpcClient};
 use pickledb::PickleDb;
 use reqwest;
+use reqwest::blocking::Client;
+use reqwest::header::USER_AGENT;
 use serde_json;
 use std::process;
+
+const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 // Use to print debug log if enabled with argument
 #[macro_export]
@@ -124,7 +128,6 @@ pub fn clear_activity(is_activity_set: &mut bool, client: &mut DiscordIpcClient)
 
 pub fn get_cover_url(
     album_id: &str,
-    last_album_id: &str,
     album: &str,
     mut _cover_url: String,
     cache_enabled: bool,
@@ -132,10 +135,6 @@ pub fn get_cover_url(
     artist: &str,
     lastfm_api_key: &str,
 ) -> String {
-    if album_id == last_album_id {
-        return _cover_url;
-    }
-
     // If no album or Unknown Album
     if album.eq("Unknown Album") {
         println!("Missing album name or Unknown Album.");
@@ -159,7 +158,12 @@ pub fn get_cover_url(
         }
     }
 
-    let request_url = format!("http://ws.audioscrobbler.com/2.0/?method=album.getinfo&api_key={}&artist={}&album={}&autocorrect=0&format=json", lastfm_api_key, url_escape::encode_component(artist), url_escape::encode_component(album));
+    let request_url = format!(
+    	"http://ws.audioscrobbler.com/2.0/?method=album.getinfo&api_key={}&artist={}&album={}&autocorrect=0&format=json",
+     	lastfm_api_key,
+     	url_escape::encode_component(artist),
+     	url_escape::encode_component(album)
+    );
 
     let mut url: String = match reqwest::blocking::get(request_url) {
         Ok(res) => match res.json::<serde_json::Value>() {
@@ -173,6 +177,101 @@ pub fn get_cover_url(
         url.pop();
         url.remove(0);
         println!("[last.fm] fetched image link: {}", url);
+
+        // Save cover url to cache
+        if cache_enabled {
+            match album_cache.set(&album_id, &url) {
+                Ok(_) => {
+                    println!("[cache] saved image url for: {}.", album_id)
+                }
+                Err(_) => {
+                    println!("[cache] error, unable to write to cache file.")
+                }
+            }
+        }
+
+        return url;
+    }
+
+    return String::from("missing-cover");
+}
+
+pub fn get_cover_url_musicbrainz(
+    album_id: &str,
+    album: &str,
+    mut _cover_url: String,
+    cache_enabled: bool,
+    album_cache: &mut PickleDb,
+    artist: &str,
+) -> String {
+    // If no album or Unknown Album
+    if album.eq("Unknown Album") {
+        println!("Missing album name or Unknown Album.");
+
+        return String::from("missing-cover");
+    }
+
+    // Load from cache if enabled
+    if cache_enabled {
+        let cache_url = if album_cache.exists(&album_id) {
+            match album_cache.get(&album_id) {
+                Some(url) => url,
+                None => String::new(),
+            }
+        } else {
+            String::new()
+        };
+
+        if (!cache_url.is_empty()) & (cache_url.len() > 5) {
+            return String::from(cache_url);
+        }
+    }
+
+    let user_agent = format!(
+        "mpris-discord-rpc/{} (patryk.kurdziel@protonmail.com)",
+        VERSION
+    );
+
+    let request_url = format!(
+    	"https://musicbrainz.org/ws/2/release/?query=artist:\"{}\"ANDrelease:\"{}\"&fmt=json&limit=1",
+    	url_escape::encode_component(artist),
+     	url_escape::encode_component(album)
+    );
+
+    let client = Client::new();
+    let mut mbid: String = match client
+        .get(request_url)
+        .header(USER_AGENT, &user_agent)
+        .send()
+    {
+        Ok(res) => match res.json::<serde_json::Value>() {
+            Ok(data) => data["releases"][0]["id"].to_string(),
+            Err(_) => String::new(),
+        },
+        Err(_) => String::new(),
+    };
+
+    if !mbid.is_empty() && (mbid.len() > 5) {
+        mbid.pop();
+        mbid.remove(0);
+    }
+
+    let mut url: String = match client
+        .get(format!("http://coverartarchive.org/release/{}/", mbid))
+        .header(USER_AGENT, &user_agent)
+        .send()
+    {
+        Ok(res) => match res.json::<serde_json::Value>() {
+            Ok(data) => data["images"][0]["thumbnails"]["small"].to_string(),
+            Err(_) => String::new(),
+        },
+        Err(_) => String::new(),
+    };
+
+    if !url.is_empty() && (url.len() > 5) {
+        url.pop();
+        url.remove(0);
+        println!("[listenbrainz] fetched image link: {}", url);
 
         // Save cover url to cache
         if cache_enabled {