+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added the ability to provide your own Last.fm API key via arguments and config file.
 - Providing a Last.fm API key during compilation is now optional.
+- Added Listenbrainz as a fallback source for album art when Last.fm doesn't provide one or API key is not set. This can be disabled with an argument or in the config.
 
 ## v0.3.0